Bit more tidying up
