+++ src/lib/machines/layout/layoutContext.ts
@@ -1,6 +1,6 @@
 import { Maybe } from "../../../types/graphql";
 import { SourcePlanRequestAction } from "../../../types/stan";
-import { LayoutPlan } from "./layoutMachine";
+import { LayoutPlan } from "./index";
 
 export interface LayoutContext {
   layoutPlan: LayoutPlan;
